# backend/src/swagger.yaml
openapi: 3.0.0
info:
  title: API du Jeu de Connexion Cinéma
  version: 1.0.0
  description: API pour le jeu de connexion entre films et acteurs américains de 1980 à 2023, utilisant la base de données movies_data-1980-2023
servers:
  - url: http://localhost:3000/api
    description: Serveur de développement
components:
  schemas:
...
    GameState:
      type: object
      required:
        - currentMovie
        - discoveredActors
        - score
        - timeSpent
        - currentLevel
        - combo
      properties:
        currentMovie:
          type: string
          description: ID du film actuel
        discoveredActors:
          type: array
          items:
            type: string
          description: Liste des IDs des acteurs découverts
        score:
          type: integer
          description: Score actuel
        timeSpent:
          type: integer
          description: Temps passé en secondes
        currentLevel:
          type: integer
          description: Niveau actuel
        combo:
          type: integer
          description: Multiplicateur de combo
        activePower:
          type: string
          description: Pouvoir actif
        activeBonus:
          type: string
          description: Bonus actif
    GameProgress:
      type: object
      properties:
        message:
          type: string
          description: Message de confirmation
        bonus:
          type: integer
          description: Bonus calculé
        totalScore:
          type: integer
          description: Score total
        powerEffect:
          type: object
          properties:
            type:
              type: string
            duration:
              type: integer
        comboBonus:
          type: object
          properties:
            type:
              type: string
            duration:
              type: integer
    WikiInfo:
      type: object
      required:
        - title
        - extract
        - imageUrl
        - fullUrl
      properties:
        title:
          type: string
          description: Titre Wikipedia
        extract:
          type: string
          description: Extrait de l'article
        imageUrl:
          type: string
          description: URL de l'image
        fullUrl:
          type: string
          description: URL de l'article complet
...
paths:
...
  /game/progress:
    post:
      summary: Sauvegarder la progression du jeu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameState'
      responses:
        '200':
          description: Progression sauvegardée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameProgress'
  /game/movies/{id}/hint:
    get:
      summary: Obtenir un indice pour un film
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Indice trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  connection_count:
                    type: integer
        '500':
          description: Erreur lors de la récupération de l'indice
  /game/movies/{id}/wiki:
    get:
      summary: Obtenir les informations Wikipedia d'un film
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Informations Wikipedia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiInfo'
        '404':
          description: Film non trouvé
  /game/actors/{id}/wiki:
    get:
      summary: Obtenir les informations Wikipedia d'un acteur
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Informations Wikipedia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiInfo'
        '404':
          description: Acteur non trouvé
...